+++ src/main/java/com/vincentdao/project_euler/Problem5.java
@@ -9,11 +9,11 @@ public class Problem5 {
     private static final long MAX = 20;
 
     public static void main(String[] args) {
-        List<Integer> primes = PrimeTableGenerator.generateWithSieveOfEratosthenes(100);
+        List<Long> primes = PrimeTableGenerator.generateWithSieveOfEratosthenes(100);
         long result = 2;
         for (long i = 3; i <= MAX; i++) {
             if (result % i != 0) {
-                for (int prime : primes) {
+                for (long prime : primes) {
                     long tmp = result * prime;
                     if (tmp % i == 0) {
                         result = tmp;
